3_e1_to_e4/requirements.txt
pydantic-ai-slim[openrouter]
pydantic
tqdm
python-dotenv
ruff
qdrant-client
openai
